refactor: share buffered output filtering
