+++ packages/wb/src/utils/output.ts
@@ -1,0 +1,18 @@
+export function shouldPrintBufferedOutput(exitCode: number, output: string): boolean {
+  return exitCode !== 0 || hasWarningOutput(removeNoColorWarning(output));
+}
+
+export function normalizeBufferedOutput(output: string): string {
+  return removeNoColorWarning(output).trim();
+}
+
+function hasWarningOutput(output: string): boolean {
+  return /\bwarn(?:ing)?s?\b/i.test(output.replaceAll(/\b0 warnings?\b/gi, ''));
+}
+
+function removeNoColorWarning(output: string): string {
+  return output.replaceAll(
+    /\(node:\d+\) Warning: The 'NO_COLOR' env is ignored due to the 'FORCE_COLOR' env being set\.\n\(Use `node --trace-warnings \.\.\.` to show where the warning was created\)\n?/g,
+    ''
+  );
+}
